+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Fixed
 - Fixed mailservers connectivity issue
+- Clear chat action correctly clear the unread messages counter
 
 ### Changed
 - Downgraded React Native to 0.53.3 for improved performance and decreased battery consumption